+++ changelog.md
@@ -22,7 +22,8 @@
 - improved icon embedding in basic.roundslider for better stylability
 - round sliders fully styled with CSS
 - reconnect interval can be configured in config.ini by the key "reconnect_time"
-- template checker recursively checks widget calls within widgets (mainly in quad widgets)
+- template checker recursively checks widget calls within widgets, e.g. dynamic icons and in quad widgets (w/ new parameter type "widget")
+- template checker now checks parameter arrays for plots and uzsu in quad widgets w/ new parameter types "plotparam" and "uzsuparam" (yet only for widgets with one plot / uzsu)
 
 ### Updated Libraries
 - MatthiasMullie/minify patched to get back to the original performance see version-info in the src folder